IGNITE-19675 [IEP-104] Implement WALIterator over ByteBuffer

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/wal/io/SimpleFileInput.java
@@ -133,8 +133,8 @@ public class SimpleFileInput extends FileInput {
     /**
      * @return Position in the stream.
      */
-    @Override public int position() {
-        return (int)(pos - buf.remaining());
+    @Override public long position() {
+        return pos - buf.remaining();
     }
 
     /**
